refactor(tests): split test_update_checks.py into three focused files
Splits the 2,261-line monolith into test_update_checks_guards.py (22 tests),
test_update_checks_prompt.py (45 tests), and test_update_checks_fetch.py
(27 tests), with shared factory helpers in _update_checks_helpers.py.
